astock_trend_backtestSkill/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0
pyarrow>=14.0.0
talib-binary>=0.4.19
statsmodels>=0.14.0
scipy>=1.11.0
optuna>=3.4.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
apscheduler>=3.10.0
pyyaml>=6.0.0
jinja2>=3.1.0
numba>=0.58.0
baostock>=0.8.8
